+++ view/frontend/web/js/view/payment/hipay-methods.js
@@ -146,11 +146,6 @@ define([
         'HiPay_FullserviceMagento/js/view/payment/method-renderer/hipay-creditlong_opc4'
     },
     {
-      type: 'hipay_klarna',
-      component:
-        'HiPay_FullserviceMagento/js/view/payment/method-renderer/hipay-klarna'
-    },
-    {
       type: 'hipay_bnpp4X',
       component:
         'HiPay_FullserviceMagento/js/view/payment/method-renderer/hipay-bnpp4X'